requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.24.1
google-generativeai==0.3.2
supabase==2.0.2
aiohttp==3.9.1
asyncio==3.4.3
datasets==2.18.0
huggingface-hub==0.21.4
tokenizers==0.15.2
librosa==0.11.0
numpy==2.2.6
pandas==2.3.1
joblib==1.5.1
scikit-learn==1.5.2 